requirements.txt
redis==2.10.6
kafka-python==1.3.5